Make the positive domain numbers unsigned
A count of something, a page, a bet, an id -- none of them can be negative, and
each said so with a validator over a signed inner type. That check cost a
`Result` at every construction and at every arithmetic operation, and it existed
only because the inner type could hold a value the domain could not. An unsigned
type makes that value unrepresentable, so the checking disappears instead of
moving somewhere else.

The width is kept, since dropping the sign bit doubles the positive range rather
than halving the need for bits. Postgres has no unsigned column, so the macro
stores such a type in the signed integer of the same width: u16 in an int2, u32
in an int4, u64 in an int8. Only u8 widens, as there is no one-byte integer.
Both directions convert instead of casting, and a refusal is only possible for a
value the column could not have held anyway.

That also unblocks u64, which had no encoding at all before and made `Position`
and `AffectedRows` Rust-only values.

What the call sites gain: `UserId::from(TeloxideUserId)` stops needing an
`expect`, and so do the battle payout and the bet-to-length conversion. Where a
value comes from a user, the refusal moves into `parse`, which is where it
belongs. `(page - 1).expect(...)` becomes `page - 1` -- saturating, which is
safer than the panic it replaces.

`TelegramChatId` stays signed: a group's id is negative, and a supergroup's
begins with -100. `AccessHash` too -- it uses the full 64-bit range.

`Offset::calculate` needed a real fix. Its comment claimed the product of two
i16 always fits an i32, which held only while both were signed; two u16 multiply
out to twice what an i32 holds. It now multiplies in i64 and cuts the result
down.

`PositiveLength` is gone -- declared, forbidden in clippy.toml, used nowhere.

+++ src/metrics.rs
@@ -644,7 +644,7 @@ impl SelfDestructionFinishedGauges {
         for state in DeletionState::TERMINAL {
             let count = counts.iter()
                 .find(|(counted, _)| *counted == state)
-                .map_or(0, |(_, count)| count.value());
+                .map_or(0, |(_, count)| count.value() as i64);
             self.0.gauge(&[&state.to_string()]).set(count);
         }
     }
